category tag search

+++ updates/create_blog_posts_table.php
@@ -10,16 +10,17 @@ class CreateBlogPostsTable extends Migration
     {
         if(PluginManager::instance()->hasPlugin('RainLab.Blog'))
         {
-            Schema::table('rainlab_blog_posts', function($table)
-            {
-                $table->string('seo_title')->nullable();
-                $table->string('seo_description')->nullable();
-                $table->string('seo_keywords')->nullable();
-                $table->string('canonical_url')->nullable();
-                $table->string('redirect_url')->nullable();
-                $table->string('robot_index')->nullable();
-                $table->string('robot_follow')->nullable();
-            });
+            if (!Schema::hasColumn('rainlab_blog_posts', 'seo_title')) {
+                Schema::table('rainlab_blog_posts', function ($table) {
+                    $table->string('seo_title')->nullable();
+                    $table->string('seo_description')->nullable();
+                    $table->string('seo_keywords')->nullable();
+                    $table->string('canonical_url')->nullable();
+                    $table->string('redirect_url')->nullable();
+                    $table->string('robot_index')->nullable();
+                    $table->string('robot_follow')->nullable();
+                });
+            }
         }
     }
 
@@ -27,18 +28,18 @@ class CreateBlogPostsTable extends Migration
     {
         if(PluginManager::instance()->hasPlugin('RainLab.Blog'))
         {
-            Schema::table('rainlab_blog_posts', function($table)
-            {
-                $table->dropColumn('seo_title');
-                $table->dropColumn('seo_description');
-                $table->dropColumn('seo_keywords');
-                $table->dropColumn('canonical_url');
-                $table->dropColumn('redirect_url');
-                $table->dropColumn('robot_index');
-                $table->dropColumn('robot_follow');
-            });
+            if (Schema::hasColumn('rainlab_blog_posts', 'seo_title')) {
+                Schema::table('rainlab_blog_posts', function ($table) {
+                    $table->dropColumn('seo_title');
+                    $table->dropColumn('seo_description');
+                    $table->dropColumn('seo_keywords');
+                    $table->dropColumn('canonical_url');
+                    $table->dropColumn('redirect_url');
+                    $table->dropColumn('robot_index');
+                    $table->dropColumn('robot_follow');
+                });
+            }
         }
-
     }
 
 }
